move ash sanctifier to own file
