+++ src/tasks/minions/monsterActivity.ts
@@ -1,84 +1,19 @@
-import { Bank, MonsterKillOptions, Monsters } from 'oldschooljs';
+import { MonsterKillOptions, Monsters } from 'oldschooljs';
 
-import { checkDegradeableItemCharges, degradeItem } from '../../lib/degradeableItems';
 import { KourendKebosDiary, userhasDiaryTier } from '../../lib/diaries';
 import { trackLoot } from '../../lib/lootTrack';
 import killableMonsters from '../../lib/minions/data/killableMonsters';
 import { addMonsterXP } from '../../lib/minions/functions';
 import announceLoot from '../../lib/minions/functions/announceLoot';
 import { prisma } from '../../lib/settings/prisma';
-import { ashes } from '../../lib/skilling/skills/prayer';
 import { SkillsEnum } from '../../lib/skilling/types';
 import { SlayerTaskUnlocksEnum } from '../../lib/slayer/slayerUnlocks';
 import { calculateSlayerPoints, getSlayerMasterOSJSbyID, getUsersCurrentSlayerInfo } from '../../lib/slayer/slayerUtil';
 import { MonsterActivityTaskOptions } from '../../lib/types/minions';
 import { roll } from '../../lib/util';
-import getOSItem from '../../lib/util/getOSItem';
+import { ashSanctifierEffect } from '../../lib/util/ashSanctifier';
 import { handleTripFinish } from '../../lib/util/handleTripFinish';
 import { makeBankImage } from '../../lib/util/makeBankImage';
-import { userStatsUpdate } from '../../mahoji/mahojiSettings';
-import { BitField } from './../../lib/constants';
-
-export async function ashSanctifierEffect(user: MUser, loot: Bank, duration: number, messages: string[]) {
-	if (!user.bank.has('Ash sanctifier')) return;
-	if (user.bitfield.includes(BitField.DisableAshSanctifier)) return;
-
-	const [hasEliteDiary] = await userhasDiaryTier(user, KourendKebosDiary.elite);
-	const ashXpModifider = hasEliteDiary ? 1 : 0.5;
-
-	let startingAshSanctifierCharges = await checkDegradeableItemCharges({
-		item: getOSItem('Ash sanctifier'),
-		user
-	});
-
-	if (startingAshSanctifierCharges === 0) return;
-
-	let chargesLeft = startingAshSanctifierCharges;
-	let totalXP = 0;
-
-	const ashesSanctified: { name: string; amount: number }[] = [];
-	for (const ash of ashes) {
-		const amount = loot.amount(ash.inputId);
-		if (amount > 0 && chargesLeft >= amount) {
-			totalXP += ash.xp * ashXpModifider * amount;
-			ashesSanctified.push({ name: ash.name, amount });
-			loot.remove(ash.inputId, amount);
-			chargesLeft -= amount;
-		} else if (amount > 0 && chargesLeft < amount) {
-			totalXP += ash.xp * ashXpModifider * chargesLeft;
-			ashesSanctified.push({ name: ash.name, amount: chargesLeft });
-			loot.remove(ash.inputId, chargesLeft);
-			chargesLeft = 0;
-			break;
-		}
-	}
-
-	if (startingAshSanctifierCharges - chargesLeft === 0) return;
-
-	await degradeItem({
-		item: getOSItem('Ash sanctifier'),
-		chargesToDegrade: startingAshSanctifierCharges - chargesLeft,
-		user
-	});
-
-	userStatsUpdate(user.id, () => ({
-		ash_sanctifier_prayer_xp: {
-			increment: Math.floor(totalXP)
-		}
-	}));
-	const xpStr = await user.addXP({
-		skillName: SkillsEnum.Prayer,
-		amount: totalXP,
-		duration,
-		minimal: true,
-		multiplier: false
-	});
-
-	const ashString = ashesSanctified.map(ash => `${ash.amount}x ${ash.name}`).join(', ');
-	messages.push(
-		`${xpStr} Prayer XP from purifying ${ashString} using the Ash Sanctifier (${chargesLeft} charges left).`
-	);
-}
 
 export const monsterTask: MinionTask = {
 	type: 'MonsterKilling',
